GetCurrency implementend

+++ CryptoTrackApp/src/models/Currency.cs
@@ -8,8 +8,8 @@ namespace CryptoTrackApp.src.models
         public int Rank {get; set;}
         public string Symbol {get; set;}
         public string Name {get; set;}
-        public int Supply {get; set;}
-        public int MaxSupply {get; set;}
+        public decimal Supply {get; set;}
+        public decimal MaxSupply {get; set;}
         public double MarketCpaUsd {get; set;}
         public double VolumeUsd24Hr {get; set;}
         public double PriceUsd {get; set;}
